dev: split modes futhermore
