+++ crates/tinymist/src/modes/mod.rs
@@ -1,0 +1,2 @@
+mod lsp;
+pub use lsp::lsp_main;
